Implemented alternative to explicitly loading testthat
Added lines of code which allow us to load and attach the testthat library
(in case the test file uses it) without having to explicitly doing so
in our source code. We do so by appending a conditional statement to
the contents of the relevant file before evaluating its metrics. This
statement loads the testthat library if installed.